some commit add ResultTestDAO

+++ src/main/java/com/netcracker/testerritto/mappers/ResultRowMapper.java
@@ -17,7 +17,7 @@ public class ResultRowMapper implements RowMapper<Result> {
         result.setDate(resultSet.getDate("result_date"));
         result.setScore(resultSet.getInt("result_score"));
         result.setStatus(ListsAttr.getvalueByid(BigInteger.valueOf(resultSet.getInt("result_status")))); //из БД получать айди , а не занчение стринга
-        result.setUserId(resultSet.getBigDecimal("user_id").toBigInteger());
+        result.setUserId(BigInteger.valueOf(resultSet.getInt("user_id")));
         return result;
     }
 }
